AABox to separate module
